+++ scripts/tf-dependencies.json
@@ -596,8 +596,8 @@
     {
         "mode": "managed",
         "type": "azurerm_cognitive_account",
-        "name": "formRecognizerAccount",
-        "module": "module.formrecognizer",
+        "name": "docIntelligenceAccountId",
+        "module": "module.aiDocIntelligence",
         "provider": "provider[\"registry.terraform.io/hashicorp/azurerm\"]",
         "instances": [
             {
@@ -613,14 +613,14 @@
         "mode": "managed",
         "type": "azurerm_key_vault_secret",
         "name": "docIntelligenceKey",
-        "module": "module.formrecognizer",
+        "module": "module.aiDocIntelligence",
         "provider": "provider[\"registry.terraform.io/hashicorp/azurerm\"]",
         "instances": [
             {
                 "dependencies": [
                     "azurerm_resource_group.rg",
                     "data.azurerm_client_config.current",
-                    "module.formrecognizer.azurerm_cognitive_account.formRecognizerAccount",
+                    "module.aiDocIntelligence.azurerm_cognitive_account.docIntelligenceAccountId",
                     "module.kvModule.azurerm_key_vault.kv",
                     "module.kvModule.data.azurerm_client_config.current",
                     "random_string.random"
@@ -674,7 +674,7 @@
                     "module.entraObjects.azuread_application.aad_mgmt_app",
                     "module.entraObjects.azuread_application_password.aad_mgmt_app_password",
                     "module.entraObjects.data.azurerm_client_config.current",
-                    "module.formrecognizer.azurerm_cognitive_account.formRecognizerAccount",
+                    "module.aiDocIntelligence.azurerm_cognitive_account.docIntelligenceAccountId",
                     "module.functions.azurerm_linux_function_app.function_app",
                     "module.functions.azurerm_service_plan.funcServicePlan",
                     "module.functions.data.azurerm_key_vault.existing_kv",
@@ -715,7 +715,7 @@
                     "module.entraObjects.azuread_application.aad_mgmt_app",
                     "module.entraObjects.azuread_application_password.aad_mgmt_app_password",
                     "module.entraObjects.data.azurerm_client_config.current",
-                    "module.formrecognizer.azurerm_cognitive_account.formRecognizerAccount",
+                    "module.aiDocIntelligence.azurerm_cognitive_account.docIntelligenceAccountId",
                     "module.functions.azurerm_service_plan.funcServicePlan",
                     "module.functions.data.azurerm_storage_account.existing_sa",
                     "module.kvModule.azurerm_key_vault.kv",
@@ -1287,7 +1287,7 @@
                     "module.entraObjects.azuread_application.aad_mgmt_app",
                     "module.entraObjects.azuread_application_password.aad_mgmt_app_password",
                     "module.entraObjects.data.azurerm_client_config.current",
-                    "module.formrecognizer.azurerm_cognitive_account.formRecognizerAccount",
+                    "module.aiDocIntelligence.azurerm_cognitive_account.docIntelligenceAccountId",
                     "module.functions.azurerm_linux_function_app.function_app",
                     "module.functions.azurerm_service_plan.funcServicePlan",
                     "module.functions.data.azurerm_storage_account.existing_sa",
